requirements.txt
torch==2.2.2
tensordict==0.3.2
packaging==24.0
torchrl==0.3.1
pettingzoo==1.24.3
pygame==2.5.2
PyYAML==6.0.1
tqdm==4.66.2
fschat[model_worker,webui]
tensorboard==2.16.2
scipy==1.13.0
pathfinding==1.0.9
flash-attn>=2.0
transformers>=4.31.0
deepspeed==0.14.1